fix: resolve build crash on /settings and api runtime warnings
