+++ app/about/page.tsx
@@ -33,6 +33,7 @@ import {
   ArrowRight,
   Workflow
 } from "lucide-react";
+export const dynamic = 'force-dynamic';
 
 export const metadata: Metadata = {
   title: "About Us | DraftDeckAI",
