+++ app/api/generate/letter/route.ts
@@ -26,14 +26,15 @@ import {
 } from "@/lib/validation";
 import { getCachedUserCredits, invalidateUserCredits } from "@/lib/cached-queries";
 
-// Service role client for credit operations
-const supabaseAdmin = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL!,
-  process.env.SUPABASE_SERVICE_ROLE_KEY!,
-);
+
 
 export async function POST(request: Request) {
   try {
+    // Service role client for credit operations
+    const supabaseAdmin = createClient(
+      process.env.NEXT_PUBLIC_SUPABASE_URL!,
+      process.env.SUPABASE_SERVICE_ROLE_KEY!,
+    );
     // ✅ AUTHENTICATION CHECK
     const authHeader = request.headers.get("Authorization");
     const token = authHeader?.replace("Bearer ", "");
@@ -120,9 +121,9 @@ export async function POST(request: Request) {
     const creditsRemaining = hasUnlimitedCredits
       ? Number.MAX_SAFE_INTEGER
       : calculateRemainingCredits(
-          userCredits.credits_total,
-          userCredits.credits_used,
-        );
+        userCredits.credits_total,
+        userCredits.credits_used,
+      );
 
     if (!hasUnlimitedCredits && creditsRemaining < creditCost) {
       return NextResponse.json(
