+++ app/api/generate/presentation-outline/route.ts
@@ -13,11 +13,6 @@ import { ACTION_COSTS, calculateRemainingCredits, hasUnlimitedDeveloperCredits }
 import { reserveCredits, refundCredits, creditReservationConflictResponse } from '@/lib/credit-operations';
 import { getCachedUserCredits, invalidateUserCredits } from '@/lib/cached-queries';
 
-const supabaseAdmin = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL!,
-  process.env.SUPABASE_SERVICE_ROLE_KEY!
-);
-
 const NEBIUS_BASE_URL =
   process.env.NEBIUS_BASE_URL ||
   'https://api.tokenfactory.us-central1.nebius.com/v1/';
@@ -56,29 +51,10 @@ Generation Requirements:
 - If this is a project/product/technical deck and page count allows, include UX/user-flow, architecture diagram, and tech stack coverage.`;
 }
 
-/**
- * Helper to create a filler slide with contextual content
- * 
- * This function is called when the AI generates fewer slides than requested,
- * padding the presentation to reach the correct slide count. It intelligently
- * determines the slide type based on position and context.
- * 
- * @param slideNumber - The 1-based index of the slide being created
- * @param totalCount - The total number of slides in the presentation
- * @param topic - The presentation topic for contextual content generation
- * @returns A slide object with type, title, content, and bullet points
- * 
- * Slide type logic:
- * - First slide (1): 'title' - Main presentation title
- * - Last slide: 'conclusion' - Summary and wrap-up
- * - Middle slide (for decks > 2 slides): 'chart' - Data visualization
- * - All others: 'content' - Regular content slides
- */
 function createFillerSlide(slideNumber: number, totalCount: number, topic: string) {
   const isFirst = slideNumber === 1;
   const isLast = slideNumber === totalCount;
 
-  // For very small decks (1-2 slides), we intentionally do not create a chart slide.
   let chartPosition: number | null = null;
   if (totalCount > 2) {
     const middlePosition = Math.ceil(totalCount / 2);
@@ -137,18 +113,18 @@ function createFillerSlide(slideNumber: number, totalCount: number, topic: strin
   };
 }
 
-// Fallback to Nebius/Qwen when Gemini fails
-const nebiusClient = new OpenAI({
-  baseURL: NEBIUS_BASE_URL,
-  apiKey: process.env.NEBIUS_API_KEY,
-});
-
 async function generateWithNebius(
   prompt: string,
   pageCount: number,
   model: NebiusOutlineModel = 'meta-llama/Meta-Llama-3.1-70B-Instruct'
 ) {
   // console.log('🔄 Using Nebius/Qwen as fallback...');
+  
+  // INITIALIZED SAFELY INSIDE THE FUNCTION
+  const nebiusClient = new OpenAI({
+    baseURL: NEBIUS_BASE_URL,
+    apiKey: process.env.NEBIUS_API_KEY || 'dummy_key_for_build', // Added dummy fallback for build time safety
+  });
   
   const completion = await nebiusClient.chat.completions.create({
     model,
@@ -181,22 +157,18 @@ Make content professional, engaging, and visually focused.`
 
   const content = completion.choices[0]?.message?.content || '[]';
   
-  // Extract JSON from response
   const jsonMatch = content.match(/\[[\s\S]*\]/);
   if (jsonMatch) {
     try {
       const parsedSlides = JSON.parse(jsonMatch[0]);
       
-      // Ensure we have the correct number of slides
       if (parsedSlides.length !== pageCount) {
         console.warn(`⚠️ Nebius generated ${parsedSlides.length} slides instead of ${pageCount}. Adjusting...`);
         
-        // If too many slides, trim to pageCount
         if (parsedSlides.length > pageCount) {
           return parsedSlides.slice(0, pageCount);
         }
         
-        // If too few slides, generate filler slides based on topic
         while (parsedSlides.length < pageCount) {
           parsedSlides.push(createFillerSlide(parsedSlides.length + 1, pageCount, prompt));
         }
@@ -208,7 +180,6 @@ Make content professional, engaging, and visually focused.`
     }
   }
   
-  // Fallback: create basic slides with exact pageCount using helper
   return Array.from({ length: pageCount }, (_, i) => {
     if (i === 0) {
       return {
@@ -226,7 +197,11 @@ Make content professional, engaging, and visually focused.`
 
 export async function POST(request: Request) {
   try {
-    // ✅ AUTHENTICATION CHECK
+    const supabaseAdmin = createClient(
+      process.env.NEXT_PUBLIC_SUPABASE_URL!,
+      process.env.SUPABASE_SERVICE_ROLE_KEY!
+    );
+
     const authHeader = request.headers.get('Authorization');
     const token = authHeader?.replace('Bearer ', '');
     
@@ -259,7 +234,6 @@ export async function POST(request: Request) {
       );
     }
 
-    // Validate pageCount to prevent invalid or abusive credit calculations
     const validatedPageCount = Number(pageCount);
     if (
       !Number.isInteger(validatedPageCount) ||
@@ -272,7 +246,6 @@ export async function POST(request: Request) {
       );
     }
 
-    // Get or create user credits (cached, 15 s TTL)
     const userCredits = await getCachedUserCredits(supabaseAdmin, user.id);
     if (!userCredits) {
       return NextResponse.json(
@@ -281,7 +254,6 @@ export async function POST(request: Request) {
       );
     }
 
-    // Check if user has enough credits - use validated page count for calculation
     const creditsPerSlide = ACTION_COSTS.presentation;
     const estimatedCreditCost = validatedPageCount * creditsPerSlide;
     const creditsRemaining = hasUnlimitedCredits
@@ -304,9 +276,6 @@ export async function POST(request: Request) {
       );
     }
 
-    // Atomically reserve the estimated credit cost BEFORE generation to
-    // prevent the TOCTOU race documented in issue #477. We refund any
-    // over-reservation later if the model returns fewer slides.
     if (!hasUnlimitedCredits) {
       const reserved = await reserveCredits(
         supabaseAdmin,
@@ -325,7 +294,6 @@ export async function POST(request: Request) {
 
     // console.log('📝 Step 1: Generating slide text content...');
 
-    // Step 1: Generate text content - Use Mistral first, then Nebius fallback
     let outlines;
     try {
       const nebiusFirst = selectedModel !== 'meta-llama/Meta-Llama-3.1-70B-Instruct';
@@ -364,7 +332,6 @@ export async function POST(request: Request) {
 
     // console.log(`✅ Generated ${outlines.length} slides`);
 
-    // If outlineOnly is requested, reconcile credits then return early.
     if (outlineOnly) {
       // console.log('🚀 Returning outline only as requested');
       if (!hasUnlimitedCredits) {
@@ -387,17 +354,14 @@ export async function POST(request: Request) {
 
     // console.log('🎨 Step 2: Generating images with FLUX AI...');
     
-    // Step 2: Generate images with FLUX (skip Mistral)
     const { generatePresentationImages } = await import('@/lib/flux-image-generator');
     const { getEnhancedImagePrompt } = await import('@/lib/presentation-styles');
     
-    // Create enhanced image prompts from slide content
     const imagePrompts = outlines.map((outline: any) => {
       const slideType = outline.type || 'content';
       const title = outline.title || '';
       const content = outline.content || outline.bulletPoints?.join(', ') || '';
       
-      // Create detailed, contextual prompt for stunning images
       let basePrompt = '';
       
       if (slideType === 'title' || slideType === 'cover') {
@@ -408,17 +372,14 @@ export async function POST(request: Request) {
         basePrompt = `Professional visual representation of "${title}", ${content.substring(0, 80)}, ${prompt}`;
       }
       
-      // Enhance with style-specific keywords
       return getEnhancedImagePrompt(basePrompt, 'modern');
     });
     
-    // Generate all images with FLUX (using 512x512 - smaller, faster)
     const imageUrls = await generatePresentationImages(imagePrompts, "512x512");
     
     // console.log(`✅ Generated ${imageUrls.length} images with FLUX`);
     // console.log('📊 Step 3: Generating chart data with Mistral AI...');
     
-    // Step 3: Generate chart data with Mistral AI (keep this for now)
     let chartDataList: any[] = [];
     try {
       chartDataList = await generateChartData(outlines, prompt);
@@ -430,7 +391,6 @@ export async function POST(request: Request) {
     
     // console.log('✨ Step 4: Combining slides with images and charts...');
     
-    // Step 4: Combine everything
     const enhancedOutlines = outlines.map((outline: any, index: number) => {
       const chartData = chartDataList.find((chart: any) => chart.slideNumber === index + 1);
       
@@ -448,8 +408,6 @@ export async function POST(request: Request) {
     // console.log('✨ Step 5: Presentation enhancement complete!');
     // console.log(`📊 Final stats: ${enhancedOutlines.length} slides, ${imageUrls.length} FLUX images, ${chartDataList.length} charts`);
     
-    // Credits were reserved at estimatedCreditCost. If the model returned
-    // fewer slides, refund the difference now and log the actual cost.
     const actualCreditCost = enhancedOutlines.length * creditsPerSlide;
     let creditsUsedAfter = userCredits.credits_used;
     if (!hasUnlimitedCredits) {
@@ -464,7 +422,6 @@ export async function POST(request: Request) {
         invalidateUserCredits(user.id);
       }
 
-      // Fire-and-forget: log write does not block the response
       supabaseAdmin
         .from('credit_usage_log')
         .insert({ user_id: user.id, action_type: 'presentation', credits_used: actualCreditCost, metadata: { pageCount: enhancedOutlines.length, prompt_length: prompt.length } })
@@ -496,4 +453,3 @@ export async function POST(request: Request) {
     );
   }
 }
-
