+++ app/api/v2/generate/resume/route.ts
@@ -1,2 +1,6 @@
-// v2 canonical URL for /api/generate/resume — re-exports current handler unchanged.
-export { dynamic, runtime, POST } from '@/app/api/generate/resume/route';
+// Define the configuration explicitly so Next.js build-time analyzer can see it
+export const dynamic = 'force-dynamic';
+export const runtime = 'nodejs';
+
+// Re-export the POST handler
+export { POST } from '@/app/api/generate/resume/route';
