+++ app/dashboard/analytics/page.tsx
@@ -1,5 +1,6 @@
 import { Metadata } from 'next';
 import AnalyticsDashboard from '@/components/dashboard/analytics-dashboard';
+export const dynamic = 'force-dynamic';
 
 export const metadata: Metadata = {
   title: "Analytics | DraftDeckAI",
