+++ app/resume/page.tsx
@@ -1,11 +1,11 @@
 "use client";
+import { Suspense, useEffect, useState } from "react";
+import { useSearchParams } from "next/navigation";
 import { SiteHeader } from "@/components/site-header";
 import { MobileResumeBuilder } from "@/components/resume/mobile-resume-builder";
 import { CreateDocumentGuard } from "@/components/ui/auth-guard";
 import { ErrorBoundary } from "@/components/ui/error-boundary";
-import { useEffect, useState, Suspense } from "react";
 import { ResumeGeneratorSkeleton } from "@/components/ui/skeleton";
-import { useSearchParams } from "next/navigation";
 
 function ResumeContent() {
   const [isLoading, setIsLoading] = useState(true);
@@ -14,34 +14,17 @@ function ResumeContent() {
   const resumeId = searchParams?.get('id') || null;
 
   useEffect(() => {
-    // Simulate loading
     const timer = setTimeout(() => setIsLoading(false), 1000);
     return () => clearTimeout(timer);
   }, []);
 
   return (
-    <CreateDocumentGuard>
-      {isLoading ? (
-        <div className="container py-8 sm:py-12 px-4 sm:px-6 lg:px-8">
-          <ResumeGeneratorSkeleton />
-        </div>
-      ) : (
-        <MobileResumeBuilder templateId={templateId} resumeId={resumeId} />
-      )}
-    </CreateDocumentGuard>
-  );
-}
-
-export default function ResumePage() {
-  return (
     <div className="min-h-screen flex flex-col relative overflow-hidden">
-      {/* Background elements matching landing page */}
       <div className="absolute inset-0 mesh-gradient opacity-20"></div>
       <div className="floating-orb w-32 h-32 sm:w-48 sm:h-48 bolt-gradient opacity-15 top-20 -left-24"></div>
       <div className="floating-orb w-24 h-24 sm:w-36 sm:h-36 bolt-gradient opacity-20 bottom-20 -right-18"></div>
       <div className="floating-orb w-40 h-40 sm:w-56 sm:h-56 bolt-gradient opacity-10 top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2"></div>
 
-      {/* Grid pattern overlay */}
       <div
         className="absolute inset-0 opacity-[0.02]"
         style={{
@@ -51,16 +34,26 @@ export default function ResumePage() {
 
       <SiteHeader />
       <main className="flex-1 relative z-10">
-        <Suspense fallback={
-          <div className="container py-8 sm:py-12 px-4 sm:px-6 lg:px-8">
-            <ResumeGeneratorSkeleton />
-          </div>
-        }>
-          <ErrorBoundary>
-            <ResumeContent />
-          </ErrorBoundary>
-        </Suspense>
+        <ErrorBoundary>
+          <CreateDocumentGuard>
+            {isLoading ? (
+              <div className="container py-8 sm:py-12 px-4 sm:px-6 lg:px-8">
+                <ResumeGeneratorSkeleton />
+              </div>
+            ) : (
+              <MobileResumeBuilder templateId={templateId} resumeId={resumeId} />
+            )}
+          </CreateDocumentGuard>
+        </ErrorBoundary>
       </main>
     </div>
+  );
+}
+
+export default function ResumePage() {
+  return (
+    <Suspense fallback={<div className="min-h-screen flex items-center justify-center">Loading...</div>}>
+      <ResumeContent />
+    </Suspense>
   );
 }
