+++ app/templates/new/page.tsx
@@ -1,28 +1,24 @@
 "use client";
+import { Suspense, useEffect, useState } from "react";
 import { SiteHeader } from "@/components/site-header";
 import { CreateTemplateForm } from "@/components/templates/create-template-form";
 import { Sparkles, FileText, Zap, Star, Wand2, Plus } from "lucide-react";
-import { useEffect, useState } from "react";
 
-// Define the page component
-export default function NewTemplatePage() {
+function NewTemplateContent() {
   const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
-    // Simulate loading
     const timer = setTimeout(() => setIsLoading(false), 1500);
     return () => clearTimeout(timer);
   }, []);
 
   return (
     <div className="min-h-screen flex flex-col relative overflow-hidden">
-      {/* Background elements matching other pages */}
       <div className="absolute inset-0 mesh-gradient opacity-20"></div>
       <div className="floating-orb w-32 h-32 sm:w-48 sm:h-48 bolt-gradient opacity-15 top-20 -left-24"></div>
       <div className="floating-orb w-24 h-24 sm:w-36 sm:h-36 bolt-gradient opacity-20 bottom-20 -right-18"></div>
       <div className="floating-orb w-40 h-40 sm:w-56 sm:h-56 bolt-gradient opacity-10 top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2"></div>
 
-      {/* Grid pattern overlay */}
       <div
         className="absolute inset-0 opacity-[0.02]"
         style={{
@@ -33,7 +29,6 @@ export default function NewTemplatePage() {
       <SiteHeader />
       <main className="flex-1 relative z-10">
         <div className="container py-8 sm:py-12 px-4 sm:px-6 lg:px-8">
-          {/* Enhanced Header */}
           <div className="text-center mb-8 sm:mb-12">
             <div className="inline-flex items-center gap-2 px-4 py-2 rounded-full glass-effect mb-4 sm:mb-6 shimmer">
               <Plus className="h-4 w-4 text-yellow-500" />
@@ -53,60 +48,36 @@ export default function NewTemplatePage() {
 
             <p className="text-base sm:text-lg lg:text-xl leading-7 sm:leading-8 text-muted-foreground max-w-2xl lg:max-w-3xl mx-auto px-4 sm:px-0">
               Build{" "}
-              <span className="font-semibold text-yellow-600">
-                professional templates
-              </span>{" "}
+              <span className="font-semibold text-yellow-600">professional templates</span>{" "}
               with our{" "}
-              <span className="font-semibold text-blue-600">
-                AI-powered creator
-              </span>
+              <span className="font-semibold text-blue-600">AI-powered creator</span>
               . Get started in seconds with{" "}
-              <span className="font-semibold bolt-gradient-text">
-                intelligent suggestions and customizable designs
-              </span>
+              <span className="font-semibold bolt-gradient-text">intelligent suggestions and customizable designs</span>
             </p>
 
-            {/* Stats bar */}
             <div className="mt-6 sm:mt-8 flex flex-wrap justify-center gap-4 sm:gap-6">
               <div className="glass-effect px-4 py-2 rounded-full hover:scale-105 transition-transform duration-300">
                 <span className="bolt-gradient-text font-bold text-sm">AI</span>
-                <span className="text-muted-foreground text-xs ml-1">
-                  Powered
-                </span>
+                <span className="text-muted-foreground text-xs ml-1">Powered</span>
               </div>
               <div className="glass-effect px-4 py-2 rounded-full hover:scale-105 transition-transform duration-300">
-                <span className="bolt-gradient-text font-bold text-sm">
-                  Smart
-                </span>
-                <span className="text-muted-foreground text-xs ml-1">
-                  Suggestions
-                </span>
+                <span className="bolt-gradient-text font-bold text-sm">Smart</span>
+                <span className="text-muted-foreground text-xs ml-1">Suggestions</span>
               </div>
               <div className="glass-effect px-4 py-2 rounded-full hover:scale-105 transition-transform duration-300">
-                <span className="bolt-gradient-text font-bold text-sm">
-                  Custom
-                </span>
-                <span className="text-muted-foreground text-xs ml-1">
-                  Design
-                </span>
+                <span className="bolt-gradient-text font-bold text-sm">Custom</span>
+                <span className="text-muted-foreground text-xs ml-1">Design</span>
               </div>
             </div>
           </div>
 
-          {/* Enhanced Form Container */}
           <div className="glass-effect p-6 sm:p-8 rounded-2xl border border-yellow-400/20 relative overflow-hidden">
-            {/* Background shimmer effect */}
             <div className="absolute inset-0 shimmer opacity-20"></div>
-
-            {/* Decorative elements */}
             <div className="absolute top-4 right-4">
               <Sparkles className="h-5 w-5 text-yellow-500 animate-pulse" />
             </div>
             <div className="absolute bottom-4 left-4">
-              <Star
-                className="h-4 w-4 text-blue-500 animate-spin"
-                style={{ animationDuration: "3s" }}
-              />
+              <Star className="h-4 w-4 text-blue-500 animate-spin" style={{ animationDuration: "3s" }} />
             </div>
 
             <div className="relative z-10">
@@ -129,16 +100,13 @@ export default function NewTemplatePage() {
             </div>
           </div>
 
-          {/* Call to action */}
           <div className="text-center mt-12 sm:mt-16">
             <div className="glass-effect p-6 sm:p-8 rounded-2xl max-w-2xl mx-auto hover:scale-105 transition-transform duration-300 relative overflow-hidden">
               <div className="absolute inset-0 shimmer opacity-30"></div>
               <div className="relative z-10">
                 <div className="flex items-center justify-center gap-2 mb-4">
                   <Zap className="h-5 w-5 text-yellow-500 animate-pulse" />
-                  <span className="font-semibold bolt-gradient-text text-lg">
-                    Ready to build your template?
-                  </span>
+                  <span className="font-semibold bolt-gradient-text text-lg">Ready to build your template?</span>
                   <Star className="h-5 w-5 text-blue-500 animate-pulse" />
                 </div>
                 <p className="text-muted-foreground text-sm mb-4">
@@ -147,15 +115,11 @@ export default function NewTemplatePage() {
                 <div className="flex flex-wrap justify-center gap-3">
                   <div className="glass-effect px-4 py-2 rounded-full text-xs hover:scale-105 transition-transform duration-300">
                     <FileText className="inline h-3 w-3 text-yellow-500 mr-1" />
-                    <span className="bolt-gradient-text font-semibold">
-                      Professional Quality
-                    </span>
+                    <span className="bolt-gradient-text font-semibold">Professional Quality</span>
                   </div>
                   <div className="glass-effect px-4 py-2 rounded-full text-xs hover:scale-105 transition-transform duration-300">
                     <Zap className="inline h-3 w-3 text-blue-500 mr-1" />
-                    <span className="bolt-gradient-text font-semibold">
-                      AI-Enhanced
-                    </span>
+                    <span className="bolt-gradient-text font-semibold">AI-Enhanced</span>
                   </div>
                 </div>
               </div>
@@ -164,5 +128,13 @@ export default function NewTemplatePage() {
         </div>
       </main>
     </div>
+  );
+}
+
+export default function NewTemplatePage() {
+  return (
+    <Suspense fallback={<div className="min-h-screen flex items-center justify-center">Loading...</div>}>
+      <NewTemplateContent />
+    </Suspense>
   );
 }
